requirements.txt
beautifulsoup4==4.10.0
google-cloud==0.34.0
google-cloud-vision==2.6.3
google-api-python-client==2.33.0
gspread==5.1.1
numpy==1.21.5
opencv-python-headless==4.5.4.60
oauth2client==4.1.3
pandas==1.3.5
redis==4.1.0
